==> 0-make_stable_beluga_summaries.ipynb <==
In this script we feed the news articles to `StableBeluge-7B` LLM in order to get the teacher summary observations that will be used throught the training phase of the lightweight BART model.

# requirements

#!pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
#!pip install transformers -q
#!pip install sentence_transformers -q
#!pip install nltk -q

import torch
from transformers import AutoModelForCausalLM, AutoTokenizer, pipeline
import json
from time import time
from sentence_transformers import SentenceTransformer
import nltk
import pandas as pd
import numpy as np
from sklearn.metrics.pairwise import cosine_similarity
from copy import deepcopy

#nltk.download('punkt')

print(torch.cuda.is_available())
print(torch.cuda.current_device())
print(torch.cuda.get_device_name(torch.cuda.current_device()))

print("Available devices ", torch.cuda.device_count())
for i in range(torch.cuda.device_count()):
    print(f'Device {i}:', torch.cuda.get_device_name(i))

print(torch.__version__)

tokenizer = AutoTokenizer.from_pretrained(
    "stabilityai/StableBeluga-7B", 
    use_fast=True
)
model = AutoModelForCausalLM.from_pretrained(
    "stabilityai/StableBeluga-7B", 
    torch_dtype=torch.bfloat16,
    #low_cpu_mem_usage=True, 
    #device_map="auto"
)#.to("cuda")
model = model.to('cuda:0')

embedding_model = SentenceTransformer("all-MiniLM-L6-v2").to('cuda:0')

torch.backends.cudnn.benchmark = True

SYS_PROMPT = "### System:\nYou are StableBeluga, an AI that follows instructions extremely well. Help as much as you can. Remember, be safe, and don't do anything illegal.\n\n"

def print_progress_bar(iteration, total, bar_length=50):
    progress = float(iteration) / float(total)
    arrow = '=' * int(round(progress * bar_length) - 1)
    spaces = ' ' * (bar_length - len(arrow))

    print(f'Progress: [{arrow + spaces}] {int(progress * 100)}%', end='\r')
    
def get_most_relevant_sentences(article_text, embeddings_model, top_n_perc=.5):
    """
    Extract only the "top_n_perc" sentences of a text "article_text".
    Top n sentences are considered the sentences that are most similar to the whole text.
    It is kind of a simple extractive summarization.
    Steps:
     1. The embedding of the whole text is computed. 
     2. The whole text is broken into individual sentences.
     3 Embeddings of each individual sentence are computed.
     4. Cosine similarity of each individual sentences against the embedding of the whole text is computed.
     5. Get top n sentences (sentences that more closely resembele the idea of the whole text).
    Receives:
     - article_text: str: The text of a news article.
     - embeddings_model: object: The model that will be used for embeddings. Model should have a .encode functionality to compute the embeddings.
    Returns:
     - Text containing only the top n % most representative sentences of a text.
    """
    # compute embedding of the whole text
    whole_text_embedding = embeddings_model.encode(article_text, show_progress_bar=False)
    # break text in sentences
    sentences = nltk.sent_tokenize(article_text)
    # store the sentences in a DataFrame
    sentences_df = pd.DataFrame(sentences, columns=['sentence'])
    # compute embeddings of each sentence individually
    sentences_embeddings = embeddings_model.encode(sentences, show_progress_bar=False)
     
    # compute cosine similarities of the whole text vs each individual sentence
    cosine_sims = cosine_similarity(
        whole_text_embedding.reshape(1, -1), 
        sentences_embeddings
    )
    # store cosine similarities on a column of the DataFrame
    sentences_df['similarity'] = cosine_sims[0]
    sentences_df.reset_index(inplace=True)
    # n sentences tied to top_n_perc of the article
    top_n = round(len(sentences)*top_n_perc)
    # Top n percent sentences that capture the main idea, sorted by how they appear in the text
    most_relevant_sentences = sentences_df.sort_values(
        # sort by similarity
        by='similarity', 
        # most similars at the top
        ascending=False
    ).head(
        # top 20%
        top_n
    ).sort_values(
        # sort them back by how they appear in the original text
        by='index'
    )[
        # get senteces
        'sentence'
    ].values.tolist() # to python list
    return ' '.join(most_relevant_sentences)

def summarize_article(article_text, model, tokenizer, top_n_perc = .2):
    """
    Summarize a single article text by using Stable Beluga
    """
    start = time()
    message = """
    
    Please summarize this:
    
    {0}
    """.format(article_text)
    prompt = f"{SYS_PROMPT}### User: {message}\n\n### Assistant:\n"
    inputs = tokenizer(prompt, return_tensors="pt").to("cuda:0")
    token_length = inputs["input_ids"].shape[1]
    if token_length > 4096:
        #print("token lengths were greater than 4096, summarizing the most relevant sentences")
        # summarize only the most importante sentences of the article
        most_relevant_pieces = get_most_relevant_sentences(
            article_text, 
            embedding_model, 
            top_n_perc = top_n_perc
        )
        summary = summarize_article(most_relevant_pieces, model, tokenizer, top_n_perc = .1)
        return summary
    with torch.no_grad():
        output = model.generate(**inputs, do_sample=False, top_p=0.95, top_k=0, max_new_tokens=150)
    response = tokenizer.decode(output[0], skip_special_tokens=False)
    end = time()
    #print(end-start)
    return response, end-start

def summarize_news_articles(news_articles_to_summarize, already_summarized_news_articles, model, tokenizer):
    """
    Trigger the batch summarization process for all news articles
    """
    total_len = len(news_articles_to_summarize)
    for i, art in enumerate(news_articles_to_summarize):
        this_art = deepcopy(art)
        output, time_taken = summarize_article(
            this_art.get('content_en'), 
            model, 
            tokenizer
        )
        assitant_ix = output.find("### Assistant:")
        summary_ix = assitant_ix+len("### Assistant:\n ")
    
        summary = output[summary_ix:]
        this_art['summary'] = summary
        already_summarized_news_articles += [this_art]
        
        if i%100 == 0:
            with open('datasets/summarized_news.json', 'w') as f:
                json.dump(already_summarized_news_articles, f)

        if i%10 == 0:
            print_progress_bar(iteration=i, total=total_len)
            torch.cuda.empty_cache()
    with open('datasets/summarized_news.json', 'w') as f:
        json.dump(already_summarized_news_articles, f)

# load the translated news articles
path_file = 'datasets/translated_news.json'
with open(path_file, 'r') as jfile:
    news_articles = json.load(jfile)

# sofar summarized news
path_file = 'datasets/summarized_news.json'
with open(path_file, 'r') as jfile:
    summarized_news_articles = json.load(jfile)

summaries_df = pd.DataFrame.from_dict(summarized_news_articles)

already_summarized = set(summaries_df['h1'])

all_news_df = pd.DataFrame.from_dict(news_articles)

all_news_titles = set(all_news_df['h1'])

pending_summary = all_news_titles - already_summarized

articles_pending_summary_df = all_news_df[all_news_df['h1'].isin(pending_summary)]

articles_pending_summary = articles_pending_summary_df.to_dict(orient='records')

len(already_summarized)

len(articles_pending_summary)

len(already_summarized)/(len(already_summarized)+len(articles_pending_summary))

summarize_news_articles(
    news_articles_to_summarize=articles_pending_summary, 
    already_summarized_news_articles=summarized_news_articles,
    model=model, 
    tokenizer=tokenizer
)

len(summarized_news_articles)

summarized_news_articles[145326]
