==> 1-make_dataset.ipynb <==
In this script we prepare the Data for the BART fine tuning task by making sure that the tokenized text and summaries dont exceed 1024 tokens

import json
import pandas as pd
import torch
import numpy as np
import time
from transformers import BartForConditionalGeneration, BartTokenizer, BartConfig
from sentence_transformers import SentenceTransformer
from torch.utils.data import DataLoader, Dataset
import nltk
from sklearn.metrics.pairwise import cosine_similarity

model_name = "facebook/bart-large"  # or another BART variant
tokenizer = BartTokenizer.from_pretrained(model_name)

embedding_model = SentenceTransformer("all-MiniLM-L6-v2").to('cuda:0')

def print_progress_bar(iteration, total, bar_length=50):
    progress = float(iteration) / float(total)
    arrow = '=' * int(round(progress * bar_length) - 1)
    spaces = ' ' * (bar_length - len(arrow))

    print(f'Progress: [{arrow + spaces}] {int(progress * 100)}%', end='\r')
    
def finish_summary(summary_str):
    summ_splitted = summary_str.split('.')
    finished_summ = '.'.join(summ_splitted[:-1])
    finished_summ += '.'
    return finished_summ

def get_most_relevant_sentences(article_text, embeddings_model, top_n_perc = .2):
    """
    Extract only the "top_n_perc" sentences of a text "article_text".
    Top n sentences are considered the sentences that are most similar to the whole text.
    It is kind of a simple extractive summarization.
    Steps:
     1. The embedding of the whole text is computed. 
     2. The whole text is broken into individual sentences.
     3 Embeddings of each individual sentence are computed.
     4. Cosine similarity of each individual sentences against the embedding of the whole text is computed.
     5. Get top n sentences (sentences that more closely resembele the idea of the whole text).
    Receives:
     - article_text: str: The text of a news article.
     - embeddings_model: object: The model that will be used for embeddings. Model should have a .encode functionality to compute the embeddings.
    Returns:
     - Text containing only the top n % most representative sentences of a text.
    """
    # compute embedding of the whole text
    whole_text_embedding = embeddings_model.encode(article_text, show_progress_bar=False)
    # break text in sentences
    sentences = nltk.sent_tokenize(article_text)
    # store the sentences in a DataFrame
    sentences_df = pd.DataFrame(sentences, columns=['sentence'])
    # compute embeddings of each sentence individually
    sentences_embeddings = embeddings_model.encode(sentences, show_progress_bar=False)
     
    # compute cosine similarities of the whole text vs each individual sentence
    cosine_sims = cosine_similarity(
        whole_text_embedding.reshape(1, -1), 
        sentences_embeddings
    )
    # store cosine similarities on a column of the DataFrame
    sentences_df['similarity'] = cosine_sims[0]
    sentences_df.reset_index(inplace=True)
    # n sentences tied to top_n_perc of the article
    top_n = round(len(sentences)*top_n_perc)
    # Top n percent sentences that capture the main idea, sorted by how they appear in the text
    most_relevant_sentences = sentences_df.sort_values(
        # sort by similarity
        by='similarity', 
        # most similars at the top
        ascending=False
    ).head(
        # top 20%
        top_n
    ).sort_values(
        # sort them back by how they appear in the original text
        by='index'
    )[
        # get senteces
        'sentence'
    ].values.tolist() # to python list
    return ' '.join(most_relevant_sentences)

# Load Data Set
DataSet consists of 295,174 news articles scrapped from a Mexican Newspaper, along with its summary. Summaries were created using `StableBeluga-7B` as the teacher. I left the LLM running for several days (weeks) in order to get all the summaries.

Relevant Columns:
- `h1`: Is the title of the news articles. Is used as key to merge with the news articles information.
- `date`: Date and time when the news articles was published.
- `author`: Name of the author who published the article.
- `content`: Article text in spanish.
- `h1_en`: Article title translated to english.
- `content_en`: Article text translated to English.
- `summary`: Summary generated by `StableBeluga-7B`

# Load the StableBeluga summaries data set
with open('datasets/summarized_news.json', "r") as f:
    summaries = json.load(f)  

summs_df = pd.DataFrame.from_dict(summaries)

summs_df.shape

summs_df.head(5)

rand_ix = np.random.randint(200000)
cont = summs_df.iloc[rand_ix]['content_en']
summ = summs_df.iloc[rand_ix]['summary']

print('Content: \n{0}'.format(cont))
print('---------')
print('Summary: \n{0}'.format(summ))

# Some summaries are rouchly finished due to the `max_new_tokens` constraint of the LLM, so for those cases, perform a gentle finish.
summs_df['complete_summary'] = summs_df['summary'].apply(lambda x: finish_summary(x))

summs_df = summs_df[['content_en', 'complete_summary']].rename(
    columns = {
        'content_en':'article', 
        'complete_summary':'summary'
    }
).copy()

# Some summaries might exceed 1024 tokens (BART-large hard limit) after tokenization. In those cases we perform a text shortening by extracting the
# top_n most relevant sentences of a text. We save the shortened text in order to use that as the true label.
# See the function `get_most_relevant_sentences` defined above for more details on how the text shortening is performed
shortened_articles = []
shortened_summaries = []
for ix, row in summs_df.iterrows():
    
    print_progress_bar(iteration=ix, total = summs_df.shape[0], bar_length=50)
    # article text
    article = row['article']
    # stable beluga summary
    summary = row['summary']
    
    # Start with the article text:
    
    # initial "top n" percent of the text to keep, in order to shorten text in case of longer than expected text
    perc_to_keep = 0.9
    while True:
        # tokenize the article text
        inputs = tokenizer(article)
        if len(inputs['input_ids'])>1024:
            # article text exceeded 1024 tokens, so shorten the article text (keep only top n %)
            article = get_most_relevant_sentences(
                article, 
                embedding_model, 
                top_n_perc = perc_to_keep
            )
            # decrease the percetage to keep by 10% in case article text still exceed 1024 tokens
            perc_to_keep -= 0.1
            # keep iterating the loop
        else:
            # article text no longer exceeds 1024 tokens
            break
    # save the article text that didn't exceed 1024 tokens
    shortened_articles.append(article)
    
    # Now for the summaries:
    
    # initial "top n" percent of the text to keep, in order to shorten text in case of longer than expected summaries
    perc_to_keep = 0.9
    while True:  
        # tokenize the summary
        targets = tokenizer(summary)
        if len(targets['input_ids'])>1024:
            # summary exceeded 1024 tokens, so shorten the summary (keep only top n %)
            summary = get_most_relevant_sentences(
                summary, 
                embedding_model, 
                top_n_perc = perc_to_keep
            )
            # decrease the percetage to keep by 10% in case summary still exceeds 1024 tokens
            perc_to_keep -= 0.1
            # keep iterating the loop
        else:
            # summary no longer exceeds 1024 tokens
            break
    # save the asummary that didn't exceed 1024 tokens
    shortened_summaries.append(summary)

summs_df['short_art'] = shortened_articles
summs_df['short_summ'] = shortened_summaries

summs_df = summs_df[['short_art', 'short_summ']].rename(
    columns = {
        'short_art':'article', 
        'short_summ':'summary'
    }
).copy()

summs_df.to_json('BART_data_set.json', orient='records')
