==> 3-evaluate.ipynb <==
import json
import pandas as pd
import torch
import numpy as np
from time import time
from transformers import BartForConditionalGeneration, BartTokenizer
from sentence_transformers import SentenceTransformer
import nltk
from sklearn.metrics.pairwise import cosine_similarity

class BARTSummarizer():
    """
    BART Summarizer model, fine tuned with "El Universal" news.
    True label summaries were generated with StableBeluga7B LLM.
    """
    def __init__(self, path_to_model = "./model_save/"):
        """
        Loads the BART pretrained model
        """
        # Load pre-trained BART model and tokenizer
        self.device = torch.device("cuda:0" if torch.cuda.is_available() else "cpu")
        model_name = "{0}bart_summarizer".format(path_to_model) 
        self.model = BartForConditionalGeneration.from_pretrained(model_name).to(self.device)
        self.tokenizer = BartTokenizer.from_pretrained(model_name)
        self.model.eval()
        self.embedding_model = SentenceTransformer("all-MiniLM-L6-v2").to(self.device)
        
    def get_most_relevant_sentences(article_text, embeddings_model, top_n_perc=.5):
        """
        Extract only the "top_n_perc" sentences of a text "article_text".
        Top n sentences are considered the sentences that are most similar to the whole text.
        It is kind of a simple extractive summarization.
        Steps:
         1. The embedding of the whole text is computed. 
         2. The whole text is broken into individual sentences.
         3 Embeddings of each individual sentence are computed.
         4. Cosine similarity of each individual sentences against the embedding of the whole text is computed.
         5. Get top n sentences (sentences that more closely resembele the idea of the whole text).
        Receives:
         - article_text: str: The text of a news article.
         - embeddings_model: object: The model that will be used for embeddings. Model should have a .encode functionality to compute the embeddings.
        Returns:
         - Text containing only the top n % most representative sentences of a text.
        """
        # compute embedding of the whole text
        whole_text_embedding = embeddings_model.encode(article_text, show_progress_bar=False)
        # break text in sentences
        sentences = nltk.sent_tokenize(article_text)
        # store the sentences in a DataFrame
        sentences_df = pd.DataFrame(sentences, columns=['sentence'])
        # compute embeddings of each sentence individually
        sentences_embeddings = embeddings_model.encode(sentences, show_progress_bar=False)
         
        # compute cosine similarities of the whole text vs each individual sentence
        cosine_sims = cosine_similarity(
            whole_text_embedding.reshape(1, -1), 
            sentences_embeddings
        )
        # store cosine similarities on a column of the DataFrame
        sentences_df['similarity'] = cosine_sims[0]
        sentences_df.reset_index(inplace=True)
        # n sentences tied to top_n_perc of the article
        top_n = round(len(sentences)*top_n_perc)
        # Top n percent sentences that capture the main idea, sorted by how they appear in the text
        most_relevant_sentences = sentences_df.sort_values(
            # sort by similarity
            by='similarity', 
            # most similars at the top
            ascending=False
        ).head(
            # top 20%
            top_n
        ).sort_values(
            # sort them back by how they appear in the original text
            by='index'
        )[
            # get senteces
            'sentence'
        ].values.tolist() # to python list
        return ' '.join(most_relevant_sentences)

    def _tokenize_article(self, article_text):
        """
        Tokenize an article text. If tokenized articles is greater than 1024 tokens,
        shorten the text, keep shortening until text no longer exceeds 1024 tokens
        """
        perc_to_keep = 0.9
        while True:
            inputs = self.tokenizer(article_text, return_tensors='pt')
            if inputs['input_ids'].shape[1]>1024:
                # tokenized article exceeded 1024 tokens
                # shorten the text
                article_text = get_most_relevant_sentences(
                    article_text, 
                    top_n_perc = perc_to_keep
                )
                perc_to_keep -= 0.1
                # keep iterating
            else:
                # text no longer exceed 1024 tokens
                break
        return inputs['input_ids'].to(self.device)
        
    def summarize(self, article_txt):
        """
        Perform the summarization task
        """
        inputs = self._tokenize_article(article_txt)
        with torch.no_grad():
            summary_ids = self.model.generate(
                inputs, 
                num_beams=4, 
                max_length=250, 
                early_stopping=True
            )
        summary = self.tokenizer.decode(
            summary_ids[0], 
            skip_special_tokens=True
        )
        return summary

# Load the StableBeluga summaries validation data set
with open('datasets/BART_validation_data.json', "r") as f:
    validation_set = json.load(f)

summarizer = BARTSummarizer()

ix = 0
art = validation_set[ix]['article']
sb_summ = validation_set[ix]['summary']
summary = summarizer.summarize(art)

print("News Article:")
print(art)
print('----------------')
print("BART Summary:")
print(summary)
print('----------------')
print("StableBeluga Summary:")
print(sb_summ)

ix = 1
art = validation_set[ix]['article']
sb_summ = validation_set[ix]['summary']
summary = summarizer.summarize(art)

print("News Article:")
print(art)
print('----------------')
print("BART Summary:")
print(summary)
print('----------------')
print("StableBeluga Summary:")
print(sb_summ)

len(validation_set)

ix = np.random.randint(0,len(validation_set))
art = validation_set[ix]['article']
sb_summ = validation_set[ix]['summary']
summary = summarizer.summarize(art)

print("News Article:")
print(art)
print('----------------')
print("BART Summary:")
print(summary)
print('----------------')
print("StableBeluga Summary:")
print(sb_summ)

def print_progress_bar(iteration, total, bar_length=50):
    progress = float(iteration) / float(total)
    arrow = '=' * int(round(progress * bar_length) - 1)
    spaces = ' ' * (bar_length - len(arrow))

    print(f'Progress: [{arrow + spaces}] {int(progress * 100)}%', end='\r')

validation_df = pd.DataFrame.from_dict(validation_set)

bart_summaries = []
for i, entry in enumerate(validation_set):
    print_progress_bar(iteration=i, total=len(validation_set))
    summary = summarizer.summarize(entry['article'])
    bart_summaries.append(summary)

validation_df['BART_summaries'] = bart_summaries

validation_df.to_json('datasets/bart_summaries_val_set.json', orient='records')

## Compare summaries by ROGUE metric

import evaluate

rouge = evaluate.load('rouge')

instruct_model_results = rouge.compute(
    predictions=bart_summaries,
    references=validation_df['summary'],
    use_aggregator=True,
    use_stemmer=True,
)

print('INSTRUCT MODEL ROGUE METRIC:')
print(instruct_model_results)

## Compare summaries using Cosine Similarity

emb_model = summarizer.embedding_model

embs1 = emb_model.encode(validation_df['summary'])
embs2 = emb_model.encode(validation_df['BART_summaries'])

# Compute cosine similarity
sims_mat = cosine_similarity(embs1, embs2)

similarities = np.diag(sims_mat)

pd.Series(similarities).describe()

The model, in average, is producing an 88% of similar summaries. On the statistics, from the minimum and the 50% statistics, we can see that there might be outliers that are dragging the average down to 88%

validation_df['summary_similarity']=similarities

# outliers
validation_df[validation_df['summary_similarity']<.1]

Outliers are comprised of empty article texts, which produce empty BART summaries

validation_df[
    (validation_df['BART_summaries']!='')&
    (validation_df['summary_similarity']<.1)
]

We can see that outliers are present, due to article texts that are comprised of empty or incoherent text, as well as baseline summaries produced by the LLM that are not good summaries. We can further refine the model by removing this noisy observations from the train data. 

For the purpose of this excercise, we can use the 50% statistic, which is `.90`, as a similarity comparison between the summaries that our lightweigt BART model produces, against the summaries produced by the `StableBeluga-7B` LLM.

# Compare the time it takes for each model to perform the Summarization task

times = []
for entry in validation_set[:500]:
    start = time()
    summary = summarizer.summarize(entry['article'])
    end = time()
    secs = end-start
    times.append(secs)

# Average time for summarization is 1.2 seconds. That is 3x faster that the 3/4 seconds that Stable Beluga takes to summarize
np.mean(times)

from transformers import AutoModelForCausalLM, AutoTokenizer, pipeline

tokenizer = AutoTokenizer.from_pretrained(
    "stabilityai/StableBeluga-7B", 
    use_fast=True
)
model = AutoModelForCausalLM.from_pretrained(
    "stabilityai/StableBeluga-7B", 
    torch_dtype=torch.bfloat16,
    #low_cpu_mem_usage=True, 
    #device_map="auto"
)#.to("cuda")
model = model.to('cuda:0')

SYS_PROMPT = "### System:\nYou are StableBeluga, an AI that follows instructions extremely well. Help as much as you can. Remember, be safe, and don't do anything illegal.\n\n"

def summarize_article(article_text, model, tokenizer, top_n_perc = .2):
    start = time()
    message = """
    
    Please summarize this:
    
    {0}
    """.format(article_text)
    prompt = f"{SYS_PROMPT}### User: {message}\n\n### Assistant:\n"
    inputs = tokenizer(prompt, return_tensors="pt").to("cuda:0")
    with torch.no_grad():
        output = model.generate(**inputs, do_sample=False, top_p=0.95, top_k=0, max_new_tokens=250)
    response = tokenizer.decode(output[0], skip_special_tokens=False)
    end = time()
    #print(end-start)
    return response, end-start

times = []
for entry in validation_set[:500]:
    summary, secs = summarize_article(entry['article'], model, tokenizer)
    times.append(secs)

# Stable Beluga taked 3.75 secons in average, to perfom a summary
np.mean(times)

The lightweigth BART model performs the summarization task 3x faster than the LLM.
